+++ rustfmt-core/rustfmt-lib/src/emitter/files.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::EmitterConfig;
+use crate::emitter::EmitterConfig;
 use std::fs;
 
 #[derive(Debug, Default)]
@@ -24,9 +24,12 @@ impl Emitter for FilesEmitter {
             original_text,
             formatted_text,
         }: FormattedFile<'_>,
-    ) -> Result<EmitterResult, io::Error> {
+    ) -> Result<EmitterResult, EmitterError> {
         // Write text directly over original file if there is a diff.
-        let filename = ensure_real_path(filename);
+        let filename = match filename {
+            FileName::Stdin => return Err(EmitterError::InvalidInputForFiles),
+            FileName::Real(path_buf) => path_buf,
+        };
         if original_text != formatted_text {
             fs::write(filename, formatted_text)?;
             if self.print_misformatted_file_names {
