+++ rustfmt-core/rustfmt-lib/src/emitter/modified_lines.rs
@@ -13,7 +13,7 @@ impl Emitter for ModifiedLinesEmitter {
             formatted_text,
             ..
         }: FormattedFile<'_>,
-    ) -> Result<EmitterResult, io::Error> {
+    ) -> Result<EmitterResult, EmitterError> {
         const CONTEXT_SIZE: usize = 0;
         let mismatch = make_diff(original_text, formatted_text, CONTEXT_SIZE);
         let has_diff = !mismatch.is_empty();
